% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_doses.R
\name{check_doses}
\alias{check_doses}
\title{Check the dosing history in the dataset.}
\usage{
check_doses(
  d,
  tool = "nonmem",
  id = "ID",
  idv = "TIME",
  dv = "DV",
  mdv = "MDV",
  evid = "EVID",
  amt = "AMT",
  rate = "RATE",
  c = "C",
  sd_sens = 1
)
}
\arguments{
\item{d}{The data file to be checked.}

\item{tool}{The pharmacometric tool the data is to be used with. Potential values are "nonmem" (the default).}

\item{id}{The unique identifier for individuals (defaults to "ID").}

\item{idv}{The independent variable (defaults to "TIME").}

\item{dv}{The dependent variable (defaults to "DV").}

\item{mdv}{Missing variable identifier (defaults to "MDV").}

\item{evid}{Event type identifier (defaults to "EVID").}

\item{amt}{The dose amount variable (defaults to "AMT").}

\item{rate}{The dose rate variable (defaults to "RATE").}

\item{c}{Record censor identifier (defaults to "C").}

\item{sd_sens}{The multiple of the standard deviation to use for identifying anomalous RATE values from the duration of infusion, obtained as AMT/RATE (defaults to 1).}
}
\value{
A list of tibbles containing problematic rows in the dataset.
}
\description{
Check the dosing history in the dataset.
}
\examples{
\dontrun{
 check_doses(nmdata)
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
