% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dupes.R
\name{check_dupes}
\alias{check_dupes}
\title{Check to see whether duplicate records exist. A wrapper around \code{\link[janitor]{get_dupes}}.}
\usage{
check_dupes(
  d,
  tool = "nonmem",
  id = "ID",
  idv = "TIME",
  dv = "DV",
  mdv = "MDV",
  evid = "EVID"
)
}
\arguments{
\item{d}{The data file to be checked.}

\item{tool}{The pharmacometric tool the data is to be used with. Potential values are "nonmem" (the default).}

\item{id}{The unique identifier for individuals (defaults to "ID").}

\item{idv}{The independent variable (defaults to "TIME").}

\item{dv}{The dependent variable (defaults to "DV").}

\item{mdv}{Missing variable identifier (defaults to "MDV").}

\item{evid}{Event type identifier (defaults to "EVID").}
}
\value{
A tibble containing duplicates.
}
\description{
Check to see whether duplicate records exist. A wrapper around \code{\link[janitor]{get_dupes}}.
}
\examples{
\dontrun{
 check_dupes(nmdata)
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
