% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_events.R
\name{check_events}
\alias{check_events}
\title{Check and summarize the numbers and types of events in the dataset.}
\usage{
check_events(d, tool = "nonmem", id = "ID", evid = "EVID", c = "C")
}
\arguments{
\item{d}{The data file to be checked.}

\item{tool}{The pharmacometric tool the data is to be used with. Potential values are "nonmem" (the default).}

\item{id}{The unique identifier for individuals (defaults to "ID").}

\item{evid}{Event type identifier (defaults to "EVID").}

\item{c}{Record censor identifier (defaults to "C").}
}
\value{
A tibble containing a summary of events in the dataset.
}
\description{
Check and summarize the numbers and types of events in the dataset.
}
\examples{
\dontrun{
 check_events(nmdata)
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
